refactor(pub): Reorder classes into packages
Remove the `utils` package and move its only class to the root. In
exchange, group the `Lockfile` and `Pubspec` classes in a new `model`
package.
