+++ plugins/package-managers/pub/src/main/kotlin/model/Lockfile.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.plugins.packagemanagers.pub
+package org.ossreviewtoolkit.plugins.packagemanagers.pub.model
 
 import com.charleskorn.kaml.Yaml
 import com.charleskorn.kaml.YamlConfiguration
@@ -39,7 +39,7 @@ import kotlinx.serialization.descriptors.SerialKind
 import kotlinx.serialization.descriptors.buildSerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 
-import org.ossreviewtoolkit.plugins.packagemanagers.pub.PackageInfo.Description
+import org.ossreviewtoolkit.plugins.packagemanagers.pub.model.PackageInfo.Description
 
 private val YAML = Yaml(configuration = YamlConfiguration(strictMode = false))
