+++ plugins/package-managers/pub/src/main/kotlin/model/Pubspec.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.plugins.packagemanagers.pub
+package org.ossreviewtoolkit.plugins.packagemanagers.pub.model
 
 import com.charleskorn.kaml.Yaml
 import com.charleskorn.kaml.YamlConfiguration
@@ -38,11 +38,11 @@ import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.serializer
 
-import org.ossreviewtoolkit.plugins.packagemanagers.pub.Pubspec.Dependency
-import org.ossreviewtoolkit.plugins.packagemanagers.pub.Pubspec.GitDependency
-import org.ossreviewtoolkit.plugins.packagemanagers.pub.Pubspec.HostedDependency
-import org.ossreviewtoolkit.plugins.packagemanagers.pub.Pubspec.PathDependency
-import org.ossreviewtoolkit.plugins.packagemanagers.pub.Pubspec.SdkDependency
+import org.ossreviewtoolkit.plugins.packagemanagers.pub.model.Pubspec.Dependency
+import org.ossreviewtoolkit.plugins.packagemanagers.pub.model.Pubspec.GitDependency
+import org.ossreviewtoolkit.plugins.packagemanagers.pub.model.Pubspec.HostedDependency
+import org.ossreviewtoolkit.plugins.packagemanagers.pub.model.Pubspec.PathDependency
+import org.ossreviewtoolkit.plugins.packagemanagers.pub.model.Pubspec.SdkDependency
 
 private val YAML = Yaml(configuration = YamlConfiguration(strictMode = false))
